*DO NOT MERGE* quick local fix
this is a quick fix to the following issue:
nodejs12 no longer supported -> upgraded to nodejs18
nodejs18 doesnt bundle aws sdk v2 automatically -> SHOULD upgrade to aws sdk v3, but this involves loads of reconfiguration -> bundle the full aws sdk v2 by removing the aws-sdk from 'externals' in webpack config.
The Problem with this fix: this requires loads of memory so you need to use the line
NODE_OPTIONS="--max-old-space-size=4096" sls deploy
to deploy. I'm not clear on the ramifications of this memory issue which is why I don't think this change should be merged